joy_box/requirements.txt
# pyenv attivato
# sudo apt install mosquitto
# sudo apt install mosquitto-clients
# pyenv global attivato python 3.8.5
# pip install pip -U
#
# MQTT client
#
paho-mqtt
gpiozero

# pi@rpi3:~ $ pinout
# .-------------------------.
# | oooooooooooooooooooo J8 |
# | 1ooooooooooooooooooo   |c
# ---+       +---+ PiZero W|s
#  sd|       |SoC|   V1.1  |i
# ---+|hdmi| +---+  usb pwr |
# `---|    |--------| |-| |-'
#
# Revision           : 9000c1
# SoC                : BCM2835
# RAM                : 512Mb
# Storage            : MicroSD
# USB ports          : 1 (excluding power)
# Ethernet ports     : 0
# Wi-fi              : True
# Bluetooth          : True
# Camera ports (CSI) : 1
# Display ports (DSI): 0
#
# J8:
#    3V3  (1) (2)  5V
#  GPIO2  (3) (4)  5V
#  GPIO3  (5) (6)  GND
#  GPIO4  (7) (8)  GPIO14
#    GND  (9) (10) GPIO15
# GPIO17 (11) (12) GPIO18
# GPIO27 (13) (14) GND
# GPIO22 (15) (16) GPIO23
#    3V3 (17) (18) GPIO24
# GPIO10 (19) (20) GND
#  GPIO9 (21) (22) GPIO25
# GPIO11 (23) (24) GPIO8
#    GND (25) (26) GPIO7
#  GPIO0 (27) (28) GPIO1
#  GPIO5 (29) (30) GND
#  GPIO6 (31) (32) GPIO12
# GPIO13 (33) (34) GND
# GPIO19 (35) (36) GPIO16
# GPIO26 (37) (38) GPIO20
#    GND (39) (40) GPIO21

